# 🐛 Fix `safeApply` never calling original `apply`