--- a/.yarn/versions/33387874.yml
+++ b/.yarn/versions/33387874.yml
@@ -1,0 +1,6 @@
+releases:
+  fast-check: patch
+
+declined:
+  - "@fast-check/ava"
+  - "@fast-check/jest"