@dubzzz 2022-08-30
Performance measurements for current main (https://github.com/dubzzz/fast-check/commit/dd5b27bc3d7c2f39b191ad4b4815f65561ab0edb) vs 3.1. vs this PR:

...